+++ src/core/struct.js
@@ -1,6 +1,10 @@
 import GObject from './gobject.js';
 import { Category,Tag } from '../lib/classes.js';
 
+function NULL_OBJECT(){
+    return Object.create(null);
+}
+
 class BuiltinCorrespond {
     from='';
     to='';
@@ -65,5 +69,6 @@ export {
     Correspond,
     Cache,
     Category,
-    Tag
+    Tag,
+    NULL_OBJECT
 }