+++ src/test.js
@@ -1,3 +1,0 @@
-import { join } from 'path';
-
-console.log(join('/','/','files'));